Parameterize LedgerDB.OnDisk over @blk@
This brings it much closer to the `LgrDB` API (we leave the in-memory datatype
more abstract).

One major advantage of this refactoring is that the _tests_ for the on disk
ledger DB are now also much simpler. The on disk ledger had its own
parameterization ("LUT": ledger under test). That was a nice idea, but we
didn't take advantage of it at all, instantiating it exactly to one ledger
only, so I've now removed that entirely and we just use TestBlock directly.
This will make further refactoring of the ledger DB easier as well.